Curse you, precedence!

+++ src/screen.c
@@ -12,14 +12,14 @@ void scr_init(void) {
 }
 
 void scr_clear(uint8_t color) {
-  uint16_t _ = color << 8 + 0x20;
+  uint16_t _ = ((uint16_t)color << 8) + 0x20;
   for (size_t i = 0; i < 2000; i++)
     buf[i] = _;
   pos = 0;
 }
 
 void scr_print(char *msg, uint8_t color) {
-  uint16_t _ = color << 8;
+  uint16_t _ = (uint16_t)color << 8;
   for (size_t i = 0; msg[i] != 0; i++)
     buf[i] = _ + msg[i];
 }
